command mapping based on server name
Until now command mapping was global, which was pretty much useless, unless the renamed commands were the same on all servers.
The change introduces a new level to the command array. The key must be a server name (first field in server array, name shown in drop-down list).

e.g.:

A server is named like this:

	$servers = [
		[ 'My awesome Redis server', '127.0.0.1', 6379 ],
	];

The command mapping for that server is set like the following:

	$command = [
		'My awesome Redis server'    => [
			'FLUSHDB'  => '',
			'FLUSHALL' => '',
			'AUTH'     => '',
			'INFO'     => '',
		],
	];